requirements.txt
dj-database-url==0.5.0
Django==2.2.6
gunicorn==20.0.4
psycopg2-binary==2.8.5
whitenoise==5.0.1
geopy==1.20.0
pbkdf2==1.3
cryptography==2.1.4
requests==2.18.4
requests-oauthlib==1.3.0
